refactor(result): 게임 결과와 관련된 클래스들을 result 패키지로 이동
