+++ src/main/java/BlackjackApplication.java
@@ -3,7 +3,7 @@ import java.util.List;
 import model.card.Card;
 import model.deck.Deck;
 import model.deck.DeckFactory;
-import model.Judge;
+import model.result.Judge;
 import view.InputView;
 import view.OutputView;
