+++ src/main/java/controller/BlackjackController.java
@@ -12,9 +12,9 @@ import model.cards.CardsFactory;
 import model.cards.DealerCards;
 import model.cards.DealerCardsFactory;
 import model.deck.Deck;
-import model.GameResult;
-import model.GameResults;
-import model.Judge;
+import model.result.GameResult;
+import model.result.GameResults;
+import model.result.Judge;
 import model.cards.PlayerCardsFactory;
 import model.Players;
 import view.InputView;
