+++ src/main/java/model/result/Judge.java
@@ -1,4 +1,4 @@
-package model;
+package model.result;
 
 import model.cards.Cards;
