+++ src/test/java/model/GameResultsTest.java
@@ -3,6 +3,8 @@ package model;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Map;
+import model.result.GameResult;
+import model.result.GameResults;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
